requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-import-export==2.0.2
django-jet==1.0.8
djangorestframework==3.11.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.3
Pillow==7.1.2
psycopg2==2.8.5
pycryptodome==3.8.2
pyotp==2.3.0
pyparsing==2.4.7
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
gunicorn==20.0.4
martor
